+++ e2e/claims.service.e2e.ts
@@ -73,6 +73,7 @@ const vcExpired = 'vcExpired';
 const electrician = 'electrician';
 const projectElectrician = 'projectElectrician';
 const projectInstaller = 'projectInstaller';
+const roleForAsset = 'roleForAsset';
 const namespace = root;
 const version = 1;
 const baseRoleDef = {
@@ -160,11 +161,16 @@ const roles: Record<string, IRoleDefinitionV2> = {
     roleName: projectInstaller,
     issuer: { issuerType: 'DID', did: [staticIssuerDID] },
   },
+  [`${roleForAsset}.${root}`]: {
+    ...baseRoleDef,
+    roleName: roleForAsset,
+    issuer: { issuerType: 'DID', did: [rootOwnerDID] },
+  },
 };
 const mockGetRoleDefinition = jest
   .fn()
-  .mockImplementation((namespace: string) => {
-    return roles[namespace];
+  .mockImplementation((roleDefNamespace: string) => {
+    return roles[roleDefNamespace];
   });
 const mockCachedDocument = jest.fn().mockImplementation((did: string) => {
   return {
@@ -337,6 +343,12 @@ describe('Сlaim tests', () => {
       data: roles[`${projectInstaller}.${root}`],
       returnSteps: false,
     });
+    await domainsService.createRole({
+      roleName: roleForAsset,
+      namespace,
+      data: roles[`${roleForAsset}.${root}`],
+      returnSteps: false,
+    });
 
     ({ didRegistry, claimsService } = await connectToDidRegistry(
       await spawnIpfsDaemon()
@@ -495,7 +507,8 @@ describe('Сlaim tests', () => {
           requestorFields,
         });
 
-        expirationTimestamp && expect(exp).toEqual(Math.floor(expirationTimestamp / 1000));
+        expirationTimestamp &&
+          expect(exp).toEqual(Math.floor(expirationTimestamp / 1000));
 
         expect(claimData).not.toContain({
           fields: [{ key: 'temperature', value: 36 }],
@@ -584,7 +597,11 @@ describe('Сlaim tests', () => {
       if (registrationTypes.includes(RegistrationTypes.OffChain)) {
         expect(issuedToken).toBeDefined();
 
-        const { claimData, signer, did } = (await didRegistry.decodeJWTToken({
+        const {
+          claimData,
+          signer,
+          did: decodedTokenDid,
+        } = (await didRegistry.decodeJWTToken({
           token: issuedToken,
         })) as { [key: string]: string };
 
@@ -596,7 +613,7 @@ describe('Сlaim tests', () => {
         });
 
         expect(signer).toBe(issuerDID);
-        expect(did).toBe(requesterDID);
+        expect(decodedTokenDid).toBe(requesterDID);
       }
 
       expect(requester).toEqual(subjectDID);
@@ -633,9 +650,11 @@ describe('Сlaim tests', () => {
     });
 
     test('asset enrollment by issuer of type DID', async () => {
-      mockGetAssetById.mockImplementationOnce(({ id }: { id: string }) => ({
-        id,
-      }));
+      mockGetAssetById.mockImplementationOnce(
+        ({ id: assetId }: { id: string }) => ({
+          assetId,
+        })
+      );
       await signerService.connect(rootOwner, ProviderType.PrivateKey);
       const assetAddress = await assetsService.registerAsset();
       const assetDID = `did:${Methods.Erc1056}:${Chain.VOLTA}:${assetAddress}`;
@@ -866,7 +885,7 @@ describe('Сlaim tests', () => {
         claim: { token: res.issuedToken },
       });
 
-      const delay = (ms) => new Promise((res) => setTimeout(res, ms));
+      const delay = (ms) => new Promise((resolve) => setTimeout(resolve, ms));
       await delay(8000);
 
       return expect(
@@ -947,6 +966,40 @@ describe('Сlaim tests', () => {
         ).toBe(true);
       });
 
+      test('should be able to issue without request and publish onchain for owned asset', async () => {
+        await signerService.connect(rootOwner, ProviderType.PrivateKey);
+        const assetClaimType = `${roleForAsset}.${root}`;
+        const assetDID = `did:${Methods.Erc1056}:${
+          Chain.VOLTA
+        }:${await assetsService.registerAsset()}`;
+
+        const claim = await issueWithoutRequest(rootOwner, {
+          subjectDID: assetDID,
+          claimType: assetClaimType,
+          registrationTypes,
+        });
+        expect(claim.onChainProof).toHaveLength(132);
+        const mockedClaim = {
+          claimType: assetClaimType,
+          isApproved: true,
+          onChainProof: claim.onChainProof,
+          claimTypeVersion: version,
+          acceptedBy: claim.acceptedBy,
+          subject: assetDID,
+        };
+        mockGetClaimsBySubject
+          .mockReset()
+          .mockImplementationOnce(() => [mockedClaim]);
+
+        await claimsService.publishPublicClaim({
+          claim: { claimType: assetClaimType },
+          registrationTypes,
+        });
+        expect(
+          await claimsService.hasOnChainRole(assetDID, assetClaimType, version)
+        ).toBe(true);
+      });
+
       test('should be able to issue without publishing onchain', async () => {
         mockGetClaimsBySubject.mockImplementationOnce(() => [role1Claim]);
 
@@ -1006,7 +1059,7 @@ describe('Сlaim tests', () => {
         const waitForRegister = new Promise((resolve) =>
           claimManager.once(
             'RoleRegistered',
-            (subject, role, version, expiry: BigNumber) =>
+            (subject, role, vers, expiry: BigNumber) =>
               resolve(expiry.toNumber())
           )
         );
@@ -1048,7 +1101,6 @@ describe('Сlaim tests', () => {
         claim: {
           claimType: `${roleName1}.${root}`,
           claimTypeVersion: version,
-          requestorFields: [],
         },
         registrationTypes,
       });
@@ -1100,12 +1152,12 @@ describe('Сlaim tests', () => {
 
     const createExampleSignedCredential = async (
       issuerFields: IssuerFields[],
-      namespace: string,
+      signedCredentialNamespace: string,
       expirationDate?: Date
     ) => {
       return await verifiableCredentialsService.createRoleVC({
         id: rootOwnerDID,
-        namespace: namespace,
+        namespace: signedCredentialNamespace,
         version: '1',
         issuerFields,
         expirationDate,
@@ -1281,7 +1333,13 @@ describe('Сlaim tests', () => {
   describe('Selfsigned claim tests', () => {
     test('Selfsigned claim should be verified', async () => {
       const claimUrl = await claimsService.createSelfSignedClaim({
-        data: { claimType: roleName1 },
+        data: {
+          profile: {
+            name: 'John Doe',
+            birthdate: '1990-01-01',
+            address: '123 Main St',
+          },
+        },
         subject: rootOwnerDID,
       });
       const claim = await didRegistry.ipfsStore.get(claimUrl);
@@ -1298,13 +1356,12 @@ describe('Сlaim tests', () => {
       const assetDID = `did:${Methods.Erc1056}:${
         Chain.VOLTA
       }:${await assetsService.registerAsset()}`;
-
       mockGetCachedOwnedAssets.mockResolvedValueOnce([
         { document: { id: assetDID }, id: assetDID },
       ]);
       const claimType = 'test claim';
       const claimUrl = await claimsService.createSelfSignedClaim({
-        data: { claimType },
+        data: { claimType, claimTypeVersion: 1 },
         subject: assetDID,
       });
       const claim = await didRegistry.ipfsStore.get(claimUrl);