+++ src/modules/claims/claims.types.ts
@@ -153,16 +153,7 @@ export interface GetClaimsByRevokerOptions {
 
 export interface CreateClaimRequestOptions {
   /** Claim request params */
-  claim: {
-    /** Role namespace */
-    claimType: string;
-
-    /** Version of the role */
-    claimTypeVersion: number;
-
-    /** Requestor fields that role is requiring */
-    requestorFields?: { key: string; value: string | number }[];
-  };
+  claim: ClaimData;
 
   /** DID of the subject */
   subject?: string;
@@ -293,7 +284,7 @@ export interface PublishPublicClaimOptions {
 
 export interface CreateSelfSignedClaimOptions {
   /** Claim data */
-  data: ClaimData;
+  data: Record<string, unknown>;
 
   /** DID of the claim subject */
   subject?: string;